+++ sperl_constant_pool.c
@@ -28,25 +28,25 @@ SPerl_CONSTANT_POOL* SPerl_CONSTANT_POOL_new(SPerl* sperl) {
   return constant_pool;
 }
 
-int32_t SPerl_CONSTANT_POOL_calculate_extend_length(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, SPerl_VALUE_T byte_size) {
+int64_t SPerl_CONSTANT_POOL_calculate_extend_length(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, SPerl_VALUE_T byte_size) {
   (void)sperl;
   (void)constant_pool;
   
-  int32_t length = (byte_size + (sizeof(SPerl_VALUE_T) - 1)) / sizeof(SPerl_VALUE_T);
+  int64_t length = (byte_size + (sizeof(SPerl_VALUE_T) - 1)) / sizeof(SPerl_VALUE_T);
   
   return length;
 }
 
-void SPerl_CONSTANT_POOL_extend(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, int32_t extend) {
+void SPerl_CONSTANT_POOL_extend(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, int64_t extend) {
   (void)sperl;
   
-  int32_t capacity = constant_pool->capacity;
+  int64_t capacity = constant_pool->capacity;
   
   if (constant_pool->length + extend >= capacity) {
     if (capacity > INT32_MAX / 2) {
       SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
     }
-    int32_t new_capacity = capacity * 2;
+    int64_t new_capacity = capacity * 2;
     constant_pool->values = (SPerl_VALUE_T*) SPerl_ALLOCATOR_UTIL_safe_realloc(constant_pool->values, new_capacity, sizeof(SPerl_VALUE_T));
     memset(constant_pool->values + capacity, 0, (new_capacity - capacity) * sizeof(SPerl_VALUE_T));
     constant_pool->capacity = new_capacity;
@@ -59,7 +59,7 @@ void SPerl_CONSTANT_POOL_push_package(SPerl* sperl, SPerl_CONSTANT_POOL* constan
   assert(sizeof(SPerl_CONSTANT_POOL_FIELD) <= sizeof(SPerl_VALUE_T));
   
   // Add package information
-  int32_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_PACKAGE));
+  int64_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_PACKAGE));
   SPerl_CONSTANT_POOL_extend(sperl, constant_pool, extend_length);
   
   // Constant pool package information
@@ -74,7 +74,7 @@ void SPerl_CONSTANT_POOL_push_sub(SPerl* sperl, SPerl_CONSTANT_POOL* constant_po
   (void)sperl;
   
   // Add sub information
-  int32_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_SUB));
+  int64_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_SUB));
   SPerl_CONSTANT_POOL_extend(sperl, constant_pool, extend_length);
   
   SPerl_CONSTANT_POOL_SUB* constant_pool_sub = (SPerl_CONSTANT_POOL_SUB*)&constant_pool->values[constant_pool->length];
@@ -101,7 +101,7 @@ void SPerl_CONSTANT_POOL_push_field(SPerl* sperl, SPerl_CONSTANT_POOL* constant_
   assert(sizeof(SPerl_CONSTANT_POOL_FIELD) <= sizeof(SPerl_VALUE_T));
   
   // Add field information
-  int32_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_FIELD));
+  int64_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, sizeof(SPerl_CONSTANT_POOL_FIELD));
   SPerl_CONSTANT_POOL_extend(sperl, constant_pool, extend_length);
 
   // Constant pool field information
@@ -149,14 +149,14 @@ void SPerl_CONSTANT_POOL_push_double(SPerl* sperl, SPerl_CONSTANT_POOL* constant
 
 void SPerl_CONSTANT_POOL_push_string(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, const char* string) {
   
-  int32_t string_length = strlen(string);
+  int64_t string_length = strlen(string);
   
   // Add string length
   SPerl_CONSTANT_POOL_extend(sperl, constant_pool, 1);
   constant_pool->values[constant_pool->length] = string_length;
   constant_pool->length++;
   
-  int32_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, string_length + 1);
+  int64_t extend_length = SPerl_CONSTANT_POOL_calculate_extend_length(sperl, constant_pool, string_length + 1);
   
   SPerl_CONSTANT_POOL_extend(sperl, constant_pool, extend_length);
