+++ sperl_field.h
@@ -8,12 +8,12 @@ struct SPerl_field {
   SPerl_OP* op_type;
   SPerl_OP* op_name;
   char* abs_name;
-  int32_t constant_pool_address;
-  int32_t abs_name_constant_pool_address;
-  int32_t package_byte_offset;
+  int64_t constant_pool_address;
+  int64_t abs_name_constant_pool_address;
+  int64_t package_byte_offset;
 };
 
 SPerl_FIELD* SPerl_FIELD_new(SPerl* sperl);
-int32_t SPerl_FIELD_get_byte_size(SPerl* sperl, SPerl_FIELD* field);
+int64_t SPerl_FIELD_get_byte_size(SPerl* sperl, SPerl_FIELD* field);
 
 #endif
